feat: support call type ring in JS client and RN Dogfood app (#74)
The call flow is designed in sync with this
(https://www.notion.so/Call-flow-ccbbd4732f6745c5ae7b10d40f1f2673)[documentation].

At the client level, the following Observables are introduced:
- `activeRingCallMeta` - This tracks the active ring call.
- `rejectedRingCallMeta` - This tracks the rejected call or the
cancelled call instance.
- ` incomingRingCalls` - All of the incoming calls are pushed inside
this array, the data is popped out when the call is accepted or rejected
and the above two states are updated accordingly.
- `acceptedRingCallDetails` - Stores the members, member user ids and
call options in the store.

The `answerCall`, `rejectCall`, `cancelCall` functions are added to
`StreamVideoClient` so that we do not expose the `sendEvent` directly.

We listen to the received events in `src/events/callUserEvents` and
update the global state accordingly.

At the RN level:
- We don't need the `getOrCreateCall` anymore.
- Loader is added when you try to join a meeting or create a ring call.
- The `call` and `activeCall` context states are removed and we utilize
the common JS store's state.


`Note: The coordinator WS connection hangs up after you try different
call event patterns. This is probably a bug with the ping pong event and
since we don't send it at in the right time the BE connection is timed
out. Please refresh the app if you face this issue for now.