feat(stats): client-side collector for WebRTC stats (#82)
This PR enables the broadcasting of WebRTC stats, plus some extra
client-side stats (latency to the connected SFU) through our state
store.

There are two kinds of stats collected:
1. Generic Stats - a collection of stats for a Peer Connection and
aggregated to provide
- Average jitter
- Average publish bitrate
- Average receiving bitrate
- Highest-resolution video track (in terms of total area - w*h)
- Potential quality limitations (in Chrome-based browsers only)

2. Participant stats - a collection of stats grouped per participant
(session_id), aggregated per stream (simulcast)
- Bytes Sent / Received
- Frame Width / Frame Height / Frames per Second
- Jitter
- Codec in use

Because of the performance impact, Participant Stats are collected on
demand (`startReportingStatsFor(sessionId)` and
`stopReportingStatsFor(sessionId)` respectively). Generic Stats are
always collected.

Open points (will be done in a follow-up PR):
- Stats reporting towards the backend needs to be moved to the LLC.
Currently, SDKs do this.

Aside from stats, as part of this PR, the TypeScript version was bumped
to v4.9.3 as it introduces more type definitions for WebRTC Stats.
Because of incompatibilities, the TS version in Angular is pinned to
v4.7.4.