feat: Finalize reference documentation scope for the client and angul… (#92)
…ar-sdk packages

# Goals

- Finalize the docs setup and the scope of the reference docs ✅
- Add meaningful doc comments to the public methods ❌ -> it feels that
our public API is still very volatile at the moment. It made no sense to
put too much energy into this. That being said, we should start adding
doc comments to new code as it has very little overhead

# Details

- Setup overview:
https://www.loom.com/share/f1c7c888904c414aae767c92a066edff
- About the doc comments:
https://www.loom.com/share/9b402e7ad24b4082b54e8fb96ee38ba2

# Open questions
- Do we want to add doc comments to internal documents as well? I think
it makes sense, even though it won't be for our integrators but for
ourselves