feat(client): Allow media stream publish before `join` is called (#192)
This PR removes the restriction where if we want to publish a media
stream before the `join` method is called, the client throws an error.
Now we don't throw an error. Instead, we wait until the join is
completed and publish the media stream.

Motivation: This hidden constraint meant that SDKs had to ensure that we
only publish a stream after `join` is called. This causes problems if we
want to implement reactive device handling: publish stream as soon as
`activeCall$` is defined.