feat(react-native): bring hidden participants to view if isSpeaking and dominant speaker true (#191)
### Goal

This PR deals with bringing dominant and speaking participants to the
visible participant's area.
The local participant always occupies the first position, as handled in
#183. The 2nd position will be taken by the dominant speaker, followed
by the speaking participants, and the rest are the non-speaking
participants. The dominant speaker still maintains their position
whether or not he/she is speaking.

This PR is built with the best possible parameters available to us by
the clients and the BE. There can be flaws in bringing participants in
view in a call of 8+ participants if everyone is speaking
simultaneously. This is planned to be dealt with by the BE using the
cascading feature. So we need to wait for them.

### Extra Side Quests
- Increased readability by creating isVideoAvailable and
isAudioAudioAvailable
- Show Avatar when video is muted or video stream not available

Co-authored-by: vanGalilea <galiliziv@gmail.com>