Call signature change (#245)
Call events are still don't work properly, so I just tested that the API
calls are sent fine in network monitor.

@khushal87 can you help with RN testing? For some reason I wasn't able
to connect with two users (one was a local deployment to my phone, the
other was a stable release from TestFlight) but I kept getting errors
that one of the users don't exist. I'm sure it was my fault somehow.