feat: Reactions (#274)
This PR introduces the initial support for call reactions. The current
list of emoji codes is quite limited and we should align it across all
SDKs but it serves the purpose of e2e testing across different
platforms.

#### Note:
The API and the component signatures aren't final and they might change
in the upcoming days/weeks.
This code is tested on MacOS and iOS only. We should check the other
OS-es (Windows, Android) as well.