feat: switch to whitelisted api_key in the token generator (#281)
### Overview

The existing `/api/auth/create-token` now requires an `api_key` to be
provided as a query parameter, next to the existing `?user_id` param.
This API key must be whitelisted on the backend before a token can be
issued.
Generated tokens have a max expiry time set to 3h from the time they
were issued. Shorter-living tokens can be generated by providing the
`?exp=<secondsFromNow>` parameter.

This endpoint is meant to be used by the sample apps we produce.