fix: storing members of call in pendingcalls and usage in RN SDK (#294)
The members were not stored in the `PendingCalls` when a new call was
created. This PR aims to solve it. As a result, `outgoingCall.users` has
the needed members who were called by the caller. The usage of avatar
and name is also solved in the incoming and outgoing call view in RN
SDK.