fix: initial and unneeded onHangupCall (#265)
Fixes an unneeded calling of this func, when the app is booted up

Calling `onHangupCall` auto. is not needed ATM in contrary to the rest.

The reason is that hanging up a call is ATM only initialized by the
participant/user itself and then we call `onHangupCall`.

In the future when permissions will be implemented and a host will be
able to end up a call for everyone this might be needed.