feat(sorting): visible participants should keep their screen position (#340)
We are tweaking our default sorting presets to skip participants who are
already visible on the screen in order to reduce the possible visual
disruption (too many participants change their "slots" on the screen).

In addition, a `CallType` and `CallTypeRegistry` type has been
introduced. With this, we and our integrators can edit/update call
configuration in a more generic fashion.