feat: implement RPC retry mechanism (#366)
In some occasions, the RPC calls we do toward the SFU can be backed off
and the SFU might ask us to retry them later.
This PR implements the desired behavior and also adds a default timeout
of 5 seconds to every RPC invocation.

In addition, an event listener leak in `Call.ts` has been fixed. Also,
in an attempt to reduce the bundle size, the `uuid` package has been
removed from the list of dependencies and replaced with a custom
implementation.