feat: add i18n (#434)
## Goal

Add internationalization package that can be re-used by any SDK in the
repo.

## Implementation details

New package `@stream-io/i18n` has been added to the repo. It is now
installed to and exposed by `video-react-sdk` and
`video-react-native-sdk`.

`StreamI18nProvider` has been added to `react-bindings. The provider
initiates the `StreamI18n` instance and propagates it to the context.
The provider is added to `StreamVideo` provider but can be used alone.

The users can override / configure:

1. language
2. provide custom translations
3. provide custom `StreamI18n` instance

The `StreamI18nContextValue` carries the translator function as well as
the instance of `StreamI18n`. The `StreamI18n` instance in turn exposes
the i18next's `i18n` instance.

### PoC

The language can now be changed in react-dogfood. Not all strings are
translated. This is just to prove, that the library works as expected +
covers the main use-cases. The same applies to react-native-dogfood app.

**React Native Note**
In order the pluralization is supported in the RN SDK, a polyfill has to
be added by the RN SDK. The documentation has been updated + the
polyfill installed to `react-native-dogfood`.

## Out of scope

1. Provide translations to all the strings