chore: Fix concurrent release problem (#498)
- Create a separate workflow for release -> only allow one concurrent
run from this workflow
- Previously, we only ran versioning for the packages that were affected
by the last commit to `main`, this is now changed to run versioning to
all packages -> in prerelease mode (which is what we are in currently),
this will make a release from each package. In release mode this will
work as expected, only changed packages will get a new release. Open GH
issue about the prerelease porblem:
https://github.com/jscutlery/semver/discussions/385