refactor(react-native): remove useCreateStreamVideoClient and dogfood app refactor to have single StreamVideo on the root (#639)
This PR focus on removing `useCreateStreamVideoClient` from the RN SDK.
After the new client changes, we can have a single instance of
`StreamVideo` in our dogfood app on the root, and the anonymous and
guest login works fine with the change. This improves the flow of the
app, as users no longer need to log in every time they change mode. App
Mode can now be changed easily.

The starter kits and cookbook are also updated to prevent build errors.

Note: The documentation will be updated once the PR is merged.