feat(react-native): add audio rooms in dogfood and related docs (#810)
NOTE: the UI is very minimal for audio room in the DF app. The scope is
to have it close to the audio room tutorial now. So don't expect proper
styling there

Also we haven't added any push notification tutorial support here

both styling in DF and push in SDK for audio room and live stream is for
later

## Preview

https://www.loom.com/share/21975c08b7bd4da39d7d005f5d4e743c