refactor(react-native): replace useCallControls hook with useMediaStreamManagement (#904)
The goal of the PR is to remove the `useCallControls` hook as
`useMediaStreamManagement` covers most of the operations of it. Also:
- the `toggleAudioMuted` and `toggleVideoMuted` are moved to the hook
`useMediaStreamManagement` for the integrator's simplicity.
- the `isVideoMuted` and `isAudioMuted` is also moved to the same hook
for the integrator's simplicity so that they don't have to take out the
mute status data from `useLocalParticipants` manually(no documentation
for this so it's easy for integrators to get lost).

PS: The relevant docs are also updated.