fix(react-native): change testing using A11y labels to testId (#901)
Since a11y labels were not used correctly. They are NOT supposed to be
keys, but more like human-friendly text. Also, a11y is a separate
project in itself and not just about labels. Let’s not ship it with
wrong and half implementation - we should drop those labels. Since the
purpose is testing, use `testId` instead.