feat(react-native): CallControls customization (#920)
This PR focuses on Call control customization. The sub-components are
exported and are mostly on par with the naming convention and pattern of
React SDK.

Integrators can use the component directly to build their own Call
controls view. We also provide a default Call controls view that can be
used if integrators do not want any customization.

---------

Co-authored-by: Santhosh Vaiyapuri <3846977+santhoshvai@users.noreply.github.com>
Co-authored-by: Santhosh Vaiyapuri <santhoshvai@gmail.com>