refactor(react-bindings): remove redundant language check (#929)
If you check the logic for `changeLanguage`, the first condition in it
says
https://github.com/GetStream/stream-video-js/blob/main/packages/i18n/src/StreamI18n.ts#L88
which internally throws the warning:
```
I18n instance is not initialized. Call yourStreamI18nInstance.init().
```

This is a yellow box warning for RN SDK.
![Simulator Screenshot - iPhone SE (3rd generation) - 2023-08-13 at 22
28
47](https://github.com/GetStream/stream-video-js/assets/39884168/bf4a2b34-dc53-4b26-8720-3b08119ef039)

-----

Now in StreamI18nContext, we have a redundant condition that seems to
`changeLanguage` if there is no `language` passed. But that condition is
used when the `i18n.isInitialized` is false/undefined, which will, in
turn, call the `this._checkIsInitialized()` from the `changeLanguage`
and return the warning.

This PR tries to solve this issue and remove the redundant condition.