feat(react-native): lobby customization (#963)
This PR introduces Lobby customization. It enables basic customization
of the controls and the join call button in the Lobby component.

Note: I had a notion that we should also be able to customize the footer
component and the part that renders the video/video fallback through a
prop but we discussed that it should be done later. Hence, its not
included in the PR. Please let me know if that needs to be added as
well.