fix: handle media device initial state when media status is undefined (#1051)
If we pass:
```
mediaDeviceInitialState={{
        initialAudioEnabled: true,
        initialVideoEnabled: true,
      }}
```
to `StreamCall` it is expected that the initial states in the lobby/call
will be unmute for the camera and microphone.

However, the mic and camera still remain muted in the Lobby. Which I
don't think should happen. The culprit here is
`call?.microphone.state.status` which is `undefined` and
`initialAudioEnabled` if `true`, `initialAudioEnabled &&
call?.microphone.state.status === 'disabled'` will always result in
`false`, no matter what.

The PR tends to fix this issue.