feat: restore remote muting functionality (#1078)
Restores the disabled Remote Muting functionality. See #1070 and #988.

### Implementation details

The determination of whether a user has been soft-muted now happens
based on event data. Previously this has been done by comparing
participant states, and it had a lot of edge cases that were hard to
handle due to the async nature of the state updates.

### Notes

While there, I took the time to update the test dependencies we use.