feat(react-native):  add lobby footer component (#1091)
In adding a meeting UI, I found I couldn't customise the text above the
button in the Lobby component. As we will be using the channelId for the
callId, I didn't like that it said they were joining this seemingly
random ID. This would likely be confusing to the user as they have no
reference to this.

Changes - 

- Add a LobbyFooter component to the Lobby and the ability to customise
it
- Updated the documentation


Tested - 

- I've tested this causes no changes in behaviour in the DogFood app and
that I customise it


Note - I've also updated the translation as there was a random capital
"J" in join. I can revert if this was intentional.