fix(react-native): incorrect starting position on reanimated floating animation (#1139)
The floating view on renimated animation was broken because the initial
drag always had start position as 0.

Upon a lot of research I found that we should not use useRef objects
inside gesture callbacks. This is because the JS variable updates are
not sent to UI thread. Instead the alternative is to do useSharedValue..
now the value is shared between JS and UI thread!

probably before we had an older gesture navigator version and it worked.
Upon some update this behaviour changed.