fix(video-demo): ensure default device is selected (#1146)
Ensure that a default video/audio device is selected.

Sometimes, the browsers may report a new `deviceId` for an already known
device. Once that happens, our previously stored device preferences
can't be respected, and we fall back on selecting the first device on the list.
This aligns with the behavior implemented in the React SDK-provided call controls.