feat(react-native): live stream components for host and viewer (#1135)
This PR focuses on introducing the LiveStream components for RN SDK.

This focuses on two main components:
- `HostLiveStream` -
https://getstream.io/video/docs/reactnative/ui-components/livestream/host-live-stream/
- `ViewerLiveStream` -
https://getstream.io/video/docs/reactnative/ui-components/livestream/viewer-live-stream/

Almost all of the subcomponents within the above two components are
customizable. The components are mainly divided into the TopView,
Layout, and Controls components.

The TopView component internally consists of `DurationBadge`,
`LiveIndicator`, and `FollowerCount` component that displays the
duration of the live stream, whether the live stream is active and the
follower count displays the number of people lives in the call.

The Controls View consists of `HostStartStreamButton`,
`ViewerLeaveStreamButton`, and the `LiveStreamMediaControls`.

The `LiveStreamLayout` component is responsible for showing the screen
sharing and the video stream of the host.

The component docs are also added.