fix(react-native-dogfood): issue in subsequent call join in livestream (#1162)
Fix the dog food app livestream flow on the viewer's end, especially
around subsequent joins. This involves the removal of the autoJoin flow
and joining the call through the button after the waiting period is
over. Moreover, on subsequent joins, the call state is checked, and if
the state is joined/joining, the lobby is simply removed.

---------

Co-authored-by: Vishal Narkhede <vishalnarkhede.iitd@gmail.com>