feat: Apply device config settings when call state becomes available (#1167)
This is a small enhancement I wanted to do for the longest time:
Previously device config settings from the backend were only applied
when we joined the call. However, sometimes we have this information
sooner in the following cases:
- `get` or `getOrCreate` is called
- `Call` is created after `queryCalls` response

This PR applies backend settings in those cases. An important thing to
note: the `mic_default_on` and `camera_default_on` are NOT applied in
the above cases, because applying those settings could start the
camera/mic, these settings are only applied after we joined the call.