feat: Fast Reconnection (#1220)
Implements Fast Reconnection flow. This should significantly speed up
the call drop recovery caused by a network switch.

### Implementation notes
In case the WebSocket connection interrupts (due to a network switch
WiFi -> 5G, or anything else), we'll attempt to re-establish a
connection to the existing SFU by reusing the existing credentials and
configuration. When this fails, we will attempt a Full Reconnect, and
involve the Coordinator in the picture, as we do it today.

### Notes
This PR is a rebase of #1212 against `main`.