feat(react-sdk): update CompositeButton layout (#1228)
Previously, CompositeButton relied on its `children` to be a button
themselves. That led to a mismatch between a hover target (the whole
CompositeButton) and a click target (e.g. an IconButton passed as
`children`).

With this update, CompositeButton renders a `button` element itself. So,
you should no longer pass IconButton or TextButton as `children` - just
pass an `Icon` or text.

This is a breaking change, since props for CompositeButton have changed.

As a bonus, most of the styling for react-dogfood's DualButton was
removed, since it just re-implemented CompositeButton with a slightly
different styling.

---------

Co-authored-by: Matvei Andrienko <matvei.andrienko@getstream.io>