feat(react-bindings): allow configuring fallback language (#1235)
Currently, `StreamI18nProvider` always sets i18next option `fallbackLng`
to false. Instead, we want to be able to configure fallback language.

That comes in handy for react-dogfood, which automatically sets default
language to user's navigator language. That means that i18n breaks in
this app if user's locale is not one of our supported ones. This is
fixed by setting fallback language set to `en`.