chore(react): Asset links update (#1238)
Update production signature and link verification for pronto. Pronto
will in theory never be accessed from the prod builds because there the
ENV switch is disabled, however since the link is present in the build
due to the ENV switcher, google tries to verify the links anyway.
Unverified links result in not working deeplinking bellow Android 11,
that is why we need also to host the prod signature on pronto in order
for the other already verified link to work bellow Android 11.