feat(react-native): add support for kotlin in expo config plugin (#1239)
Partially resolves #1231 

Hello,

according to our conversation in #1231 , I'm contributing a fix to add
support for Kotlin inside MainActivity which ensures that this config
plugin works with Expo SDK 50 and React Native 0.73. I already tested
this change with Expo SDK 50 Beta and it worked fine for me, except for
instantiating the `StreamVideoClient`, but that's an unrelated issue.

Let me know if anything else needs to be adjusted.