fix: commonjs optional libs bug workaround (#1292)
In Metro, while  JS is being used there is a buggy behaviour

Suppose we have the following code

```
try {
  messaginglib = require('@react-native-firebase/messaging');
} catch (_e) {}
try {
  expoTaskManagerLib = require('expo-task-manager');
} catch (_e) {}
```

Say if messageLib is not installed and expo-task-manager is installed.

Once the error for messagingLib is catched. Expo-task-manager cannot be
imported too. This occurs only in JS code.


The workaround is to split them into different files

## Extra

Update tsconfig with the output from
https://www.npmjs.com/package/create-react-native-module