fix(publisher): ensure initial bitrate is set (#1322)
Fixes an issue where the track's bitrate is announced as 0 but in reality it isn't.

Sometimes, the track settings aren't initialized on time, and hence, the
track dimensions are reported as 0x0 by the browser.
Naturally, we expect 0 bits for encoding these track dimensions.
However, once the browser recovers and real track dimensions can be
determined, we have already sent the track announcement.
Now, instead of announcing 0 bitrate, we report some default values for
each simulcasted track. The actual value would be adjusted by the SFU
once the Publisher peer connection begins sending the video tracks.