fix(state): optimized Call State updates (#1330)
- Observables emitting primitive values will emit only when the value changes
`false` -> `false` - the subscribers won't run for the second update as the value didn't change

- updateOrAddParticipant doesn't iterate the participant array twice before updating or adding a participant
We were iterating the participant array to figure out whether to add a participant or update an existing one. Then, we create a shallow clone of the array and apply the updates. This isn't optimal.
Now, these two operations are merged into one operation.