fix(android):  no ringtone when incoming call was through foreground service (#1402)
Sound has to be set on channel level for Android 8 and above on creation
and cannot be updated later. For Android 7 and below, the sound should
be set on the notification level.

Since we relied on the notification level while moving to use the same
foreground services, the sound broke if the foreground service channel
was already created.