fix: perform full reconnect if ice restart fails (#1408)
When networks change, we perform an ICE restart to re-establish connection to the SFU.
However, when this process fails, the call stays in limbo mode (no tracks are flowing anymore). This is manifested by black tracks on the client's screens.

This PR updates the behavior of the SDK and now, we will perform full reconnect eagerly.
In the future, once we implement the Reconnect v2, this issue will be solved more elegantly.