fix: handle session_participant_count_updated event (#1467)
Handlers for the new `call.session_participant_count_updated` event. 
The Coordinator will occasionally deliver this event to all clients, and it will carry the most recent state of:
- `anonymous_participant_count` and
- `participant_count_by_role`

Specs:
https://www.notion.so/stream-wiki/Participant-counts-4411743f0b9345c192b243ccc25099a5

### Implementation notes
The `CallSession` is extended with one extra property `anonymous_participant_count`.
This value and the aggregated value for the regular participant count will be used for populating the `call.state.anonymousParticipantCount` and `call.state.participantCount` accordingly but only when the `callingState` isn't `JOINED`. Once the call is JOINED, these values will be populated from the SFU heart checks.

This enables our `useParticipantCount()` and `useAnonymousParticipantCount()` hooks to work as expected, regardless if the user has joined the call or not.