fix: don't attempt to recover broken WebSockets when there isn't a network connection (#1490)
We don't attempt to recover broken WebSockets when there isn't a network
connection. Instead, we wait for connectivity to be restored first.

This behavior created some weird bugs with `call.leave()`. As we were
creating new WebSockets, the `signalReady` never resolved (the socket
couldn't be opened) and hence, `send()` was blocking forever.
Due to this behavior, when attempting to leave a call while offline, the
`callingState` never transitioned from `OFFLINE -> LEFT`.

Ref:
https://getstream.slack.com/archives/C06H93HC60Z/p1725636744842199?thread_ts=1724334671.814709&cid=C06H93HC60Z

---------

Co-authored-by: Santhosh Vaiyapuri <santhoshvai@gmail.com>