fix: do not always error out api calls when web socket initially failed (#1495)
## Problem

Currently for all axios requests, we wait for `connectionIdPromise`.
This promise is rejected or resolved based on the first WS connect
attempt. If subsequent reconnects were successful but the first connect
was rejected, this promise would still stay rejected.

## Solution

To allow for reconnects to resolve, say for example, in 3rd try the
connection got open, we can wait for the `connectionOpen` promise to
resolve. To allow for 2 or 3 retries, we use `_waitForHealthy` function
here. Alternatively we could await only on `connectionOpen` promise
without allowing for timeouts to wait for reconnections also.