feat: add a prop to control mirroring of local participant video (#1506)
We get a lot of questions about disabling mirroring of local participant
video. It's currently possible to do by overriding the
`str-video__video--mirror` CSS class, but it seems that things will be
easier if we have a prop to control this behavior.

This PR adds an optional `mirrorLocalParticipantVideo` prop to all
built-in layout components. Setting it to `false` disable local
participant video mirroring.