fix: check for user capabilities before rendering call control buttons (#1513)
Adds OwnCapability checks on the Call Control buttons.
They shouldn't be rendered at all when certain user capabilities are not present.

## Fixes
When browser permissions are not granted, and the user doesn't have
`send-video` or `send-audio` permissions, the `CallControls` component
is still prompting for browser permissions.
The reason for this prompt is - these buttons use `useMicrophoneState()`
and `useCameraState()` hooks that trigger these prompts. Fixing this in
the buttons component will require a bit of refactoring, so I'm opting
for this less ideal and simple fix.