fix: ignore camera direction for desktop devices (#1521)
Our default call types include a default setting for camera direction
(user-facing). Having this setting applied by default is useful for
mobile devices, but on desktop devices where camera direction reporting
is not reliable it leads to confusion, because it can cause a device
other than the system default to be used by default.