fix: make device selection by device id exact (#1538)
Somewhere around Chrome v126 a new behavior was introduced, and the
`deviceId: string` constraint when querying `getUserMedia` was no longer
respected. The default device was always returned instead.

Using the `deviceId: { exact: string }` constraint seems to solve the
issue. However, it also can cause the `OvercontrainedError` exception
when device id is not valid. We handle it by falling back to default
device in this case.